requirements.txt
git+https://github.com/django-nonrel/django
Markdown==2.5.2
django-bower==5.0.1
django-extensions==1.4.8
django-filter==0.9.1
git+https://github.com/django-nonrel/mongodb-engine
django-toolbelt==0.0.1
djangorestframework==3.0.0
git+https://github.com/django-nonrel/djangotoolbox
gunicorn==19.1.1
pymongo==2.7.2
python-dateutil==2.3
python-mimeparse==0.1.4
requests==2.5.0
six==1.8.0
static3==0.5.1
whitenoise==1.0.5
wsgiref==0.1.2